pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distfl-client"
version = "1.0.0"
description = "Production-grade Python Client SDK for room-based Federated Learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DistFL Team"},
]
keywords = ["federated-learning", "machine-learning", "distributed", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
distfl = "fl_client.cli.main:main"
fl-client = "fl_client.cli.main:main"   # legacy alias

[tool.setuptools.packages.find]
include = ["fl_client*"]
exclude = ["tests*", "fl_client.web.ui*", "fl_client.dashboard*"]

[tool.setuptools.package-data]
fl_client = [
    "web/static/**/*.html",
    "web/static/**/*.js",
    "web/static/**/*.css",
    "web/static/**/*.svg",
    "web/static/**/*.png",
    "web/static/**/*.ico",
    "web/static/**/*.json",
    "web/static/**/*.woff",
    "web/static/**/*.woff2",
    "web/static/**/*.ttf",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
